Create a master thesis.tex file
This is the main component of the thesis, basically setting all the
latex options and then including each of the components. I have
used latex for this part since it is a document primarily consisting of
latex commands and this avoids having to deal with the conversion of
this file between filetypes.