Manually include sections in thesis.tex
Rather than have some code automatically pull in all the generated
files, I am going to do so manually. This gives more contol over the
structure of the document and I am not going to have so many files that
this is unmanageable. It also allows me to include the \chapter commands
and use multiple files per chapter.