feat: Include projects as submodules
By having each of the projects comprising my thesis as submodules, I am
able to use these paths for the graphics search path meaning that I just
need to update the figure in the project and it will be updated in the
thesis. Same this goes for code snippets or anything else.

This does mean a maintenance burden of having to update the submodules
however there will be no issues of am I using the latest version of this
figure.