chore: Makefile rule to convert svg files to pdf
The figures output from altair are in the svg format which is not able
to be included in the latex document. However there is the handy
cariopdf package which allows me to convert the svg to a pdf file. This
creates a makefile rule to convert all svg files to a pdf so they can be
included in the document. Part of this change includes updating the
environment to include the necessary executables for the conversion.