fix: Use input instead of include for adding files
This adds files to thesis.tex using input, rather than include. This
change has been made to deal with some issues in formatting, most
notably the clearing of pages. Using include was like each page was
compiled separately and I was introducing additional issues by using it.
Input on the other hand is making it like there is a single large
document, which is kinda what I want.