writing(melting): Document packing algorithm
This quickly documents the packing algorithm in the methods of the
Crystal Melting section. This is probably the most appropriate place
for this.

This change removes the planning of the packing algorithm from the
methods chapter, where it didn't really rate a section for what I was
going to write.

Fixes #35