chore: Configure pre-commit to use markdownlint
This uses the same markdownlint I am using in the CI as a pre-commit
hook. Recently I have had a number of issues with the CI picking up
stylistic issues which are then frustrating to fix. To make sure that
the text which is uploaded doesn't contain the same issues I am
configuring it in pre-commit