planning(ML): Background info in introduction
This moves all of the background information for the machine learning
chapter to the introduction. There was a reasonable amount of
information in the methods section, however it didn't really belong
there so I have moved it out.

Also the methods makes use of the method section in the crystal melting
chapter to describe the simulations.