editing(melting): Trying to structure phase transition
The structuring of the phase transition is turning out to be more
complicated than I was hoping. There are a lot of ideas which I am
presenting and I am finding it hard to bring them all together.

This is my first attempt at editing this section which will need some
more work.