fix(structure): Put methods chapter after introduction
The methods chapter includes many of the ideas that are used throughout
the rest of the thesis, including the computational tools used to
generate  the results. While the approach I took was based on the
structure of a Nature Materials paper, the thesis examiners thought the
methods was better placed towards the start of the thesis.

Ref #229